requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.2
dnspython==1.16.0
Flask==3.1.1
frozenlist==1.7.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
multidict==6.6.3
propcache==0.3.2
pymongo==4.13.2
python-dotenv==1.1.1
typing_extensions==4.14.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0